aucd/pyproject.toml
[tool.poetry]
name = "auc"
version = "2.0.0"
description = "A tool for checking for pacman based distro updates. It's probably really bad."
authors = ["Lucy Hancock <lucy@leh.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pyyaml = "^6.0.2"
requests-unixsocket2 = "^0.4.2"
urllib3 = "^2.2.3"

[tool.poetry.scripts]
auc = "auc.cli.main:run"
aucd = "auc.daemon.main:run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
